hoik up .lms to PointCloud for broader backwards-compat
